change how the Concentration Check item is made
Instead of using an exported item as the base and updating it, start with a basic itemData with the setting we need then use Item.create to use the item template to help flush out the item. It sets temporary to true to make sure it's not actually persisted in the database.